Define GitLab CI jobs

Define GitLab CI jobs

gl_ci_job(job_name, stage = job_name, allowed_dependencies = c(), ...)

gl_default_ci_pipeline()

use_gitlab_ci(
  pipeline = gl_default_ci_pipeline(),
  image = "rocker/r-devel:latest",
  push_to_remotes = c(),
  path = ".gitlab-ci.yml",
  overwrite = TRUE,
  add_to_Rbuildignore = TRUE
)

Arguments

job_name
Name of job template to get CI definition elements

stage
Name of stage job belongs to

allowed_dependencies
List of job names that are allowed to be listed as dependencies of jobs. Usually this is all existing other jobs.

...
passed on to ci_r_script: booleans vanilla or slave translate to R executable options with the same name

pipeline
a CI pipeline defined as a list of lists

image
Docker image to use in GitLab ci. If NULL, not specified!

push_to_remotes
named list of remotes the code should be pushed to. Only master is pushed and for every remote a job of stage "push" is generated. See example for how to use credentials from environment variables.

path
destination path for writing GitLab CI yml file

overwrite
whether to overwrite existing GitLab CI yml file

add_to_Rbuildignore
add CI yml file (from path) to .Rbuildignore?

Examples

gl_ci_job("build", allowed_dependencies = "test")

#> $stage
#> [1] "build"
#> 
#> $script
#> $script[[1]]
#> [1] "R --vanilla -e 'library(devtools); devtools::build(path = \"./\")'"
#> 
#> 
#> $artifacts
#> $artifacts$paths
#> $artifacts$paths[[1]]
#> [1] "*.tar.gz"
#> 
#> 
#> 

use_gitlab_ci(image = "pointsofinterest/gitlabr:latest", path = tempfile(fileext = ".yml"))
use_gitlab_ci(image = "pointsofinterest/gitlabr:latest", , path = tempfile(fileext = ".yml"),
 push_to_remotes = list("github" =
 "https://${GITHUB_USERNAME}:${GITHUB_PASSWORD}@github.com/jirkalewandowski/gitlabr.git"))