Shiny module to login to GitLab API

The UI contains a login and a password field as well as an (optional) login button. The server side function returns a reactive GitLab connection, just as gl_connection() and gl_project_connection().

glLoginInput(id, login_button = TRUE)

glReactiveLogin(
  input,
  output,
  session,
  gitlab_url,
  project = NULL,
  api_version = 4,
  success_message = "GitLab login successful!",
  failure_message = "GitLab login failed!",
  on_error = function(...) {     stop(failure_message) }
)

Arguments

id
shiny namespace for the login module

login_button
whether to show a login button (TRUE) or be purely reactive (FALSE)

input
from shinyServer function, usually not user provided

output
from shinyServer function, usually not user provided

session
from shinyServer function, usually not user provided

gitlab_url
root URL of GitLab instance to login to

project
if not NULL, a codegl_project_connection is created to this project

api_version
A character with value either "3" or "4" to specify the API version that should be used

success_message
message text to be displayed in the UI on successful login

failure_message
message text to be displayed in the UI on login failure in addition to HTTP status

on_error
function to be returned instead of GitLab connection in case of login failure

Value

An input or output element for use in shiny UI.

Details

glLoginInput is supposed to be used inside a shinyUI, while glReactiveLogin is supposed to be passed on to shiny::callModule()