Add .gitlab-ci.yml file in your current project from template

Add .gitlab-ci.yml file in your current project from template

use_gitlab_ci(
  image = "rocker/verse:latest",
  repo_name = "https://packagemanager.rstudio.com/all/__linux__/focal/latest",
  url = "https://gitlab.com",
  path = ".gitlab-ci.yml",
  overwrite = TRUE,
  add_to_Rbuildignore = TRUE,
  type = "check-coverage-pkgdown"
)

Arguments

image
Docker image to use in GitLab ci. If NULL, not specified!

repo_name
REPO_NAME environment variable for R CRAN mirror used

url
url of the GitLab instance

path
destination path for writing GitLab CI yml file

overwrite
whether to overwrite existing GitLab CI yml file

add_to_Rbuildignore
add CI yml file (from path) to .Rbuildignore?

type
type of the CI template to use

Value

Used for side effects. Creates a .gitlab-ci.yml file in your directory.

Details

Types available are:

"check-coverage-pkgdown": Check package along with Code coverage with covr and pkgdown site on GitLab Pages

"check-coverage-pkgdown-renv": Check package built in a fixed renv state along with Code coverage with covr and pkgdown site on GitLab Pages.

"bookdown": Build bookdown HTML and PDF site on GitLab Pages

"bookdown-production": Build bookdown HTML and PDF site on GitLab Pages. Where principal page is for branch named 'production' and "dev/" sub-folder is for 'main' (or 'master') branch.

Examples

# Create in another directory
use_gitlab_ci(image = "rocker/verse:latest", path = tempfile(fileext = ".yml"))
if (FALSE) {
# Create in your current project with template for packages checking
use_gitlab_ci(image = "rocker/verse:latest", type = "check-coverage-pkgdown")
}